pyproject.toml
[project]
name = "SpiderX"
version = "0.1.0"
description = "A blasting login penetration testing tool that uses web crawler technology to automatically bypass front-end JS encryption"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "ddddocr>=1.5.6",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "selenium>=4.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.uv.pip]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[tool.uv.workspace]
members = [
    "playground",
]
